backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pypdf>=4.2
openai>=1.40
python-dotenv>=1.0